requirements.txt
apache-airflow
pendulum
lxml
beautifulsoup4
apache-airflow-providers-mysql